// src/app/shared/models/categoryListing.model.ts
// export class CategoryListing {
//   title: string;
//   subtitle?: string;
//   iconImgSrc?: string;

//   constructor(title: string, subtitle?: string, iconImgSrc?: string) {
//     this.title = title;
//     this.subtitle = subtitle;
//     this.iconImgSrc = iconImgSrc;
//   }

// }

export interface CategoryListing {
  title: string;
  subtitle?: string;
  iconImgSrc?: string;
}